change int to float

+++ domain/model/entities/spacecraft.go
@@ -8,12 +8,12 @@ type SpacecraftEntities struct {
 	TonsOfThrust      float64
 	TonsOfWeight      float64
 	TypeFuel          string
-	Height            int
+	Height            float64
 	Country           string
 	Objective         string
 	Activated         bool
 	ExpirationDate    time.Time
-	PeopleCapacity    int
+	PeopleCapacity    float64
 	TransportCapacity float64
 	MassiveSpeed      float64
 	PropulsionSystem  string
