+++ infrastructure/drivenadapters/adapters/entityData/spacecraft.go
@@ -8,12 +8,12 @@ type SpacecraftData struct {
 	TonsOfThrust      float64   `bson:"tonsOfThrust"`
 	TonsOfWeight      float64   `bson:"tonsOfWeight"`
 	TypeFuel          string    `bson:"typeFuel"`
-	Height            int       `bson:"height"`
+	Height            float64   `bson:"height"`
 	Country           string    `bson:"country"`
 	Objective         string    `bson:"objective"`
 	Activated         bool      `bson:"activated"`
 	ExpirationDate    time.Time `bson:"expirationDate"`
-	PeopleCapacity    int       `bson:"peopleCapacity"`
+	PeopleCapacity    float64   `bson:"peopleCapacity"`
 	TransportCapacity float64   `bson:"transportCapacity"`
 	MassiveSpeed      float64   `bson:"massiveSpeed"`
 	PropulsionSystem  string    `bson:"propulsionSystem"`
